refactor: add component folder on main page
