+++ react-project/src/pages/main-page/components/searchLogo.tsx
